% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{coef.power_model}
\alias{coef.power_model}
\title{coef method for power_model objects}
\usage{
\method{coef}{power_model}(object, transform = object$transform, ...)
}
\arguments{
\item{object}{A 'power_model' object.}

\item{transform}{Logical, whether to back-transform from log-phi.
Defaults to object$transform.}

\item{...}{Unused.}
}
\value{
Named numeric vector of coefficients (beta0, beta1, beta2).
}
\description{
coef method for power_model objects
}
