% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{fit_model}
\alias{fit_model}
\title{Fit parametric power model}
\usage{
fit_model(
  dat,
  alpha = 0.05,
  test = "onesided",
  start = c(logbeta0 = 0, logtheta = 0, logr = 0)
)
}
\arguments{
\item{dat}{Data frame with y, m1, m2.}

\item{alpha}{Significance level.}

\item{test}{'onesided' or 'twosided'.}

\item{start}{Named numeric vector of log-parameters.}
}
\value{
An object of class 'power_model'.
}
\description{
Fit parametric power model
}
