% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{logLik.power_model}
\alias{logLik.power_model}
\title{logLik method for power_model objects}
\usage{
\method{logLik}{power_model}(object, ...)
}
\arguments{
\item{object}{A 'power_model' object.}

\item{...}{Further arguments (ignored).}
}
\value{
An object of class "logLik".
}
\description{
logLik method for power_model objects
}
