% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{model_data.power_model}
\alias{model_data.power_model}
\title{Extract model data from a power_model}
\usage{
model_data.power_model(object)
}
\arguments{
\item{object}{A 'power_model' object.}
}
\value{
A data.frame with columns y, m1, m2.
}
\description{
Extract model data from a power_model
}
