% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{nobs.power_model}
\alias{nobs.power_model}
\title{nobs method for power_model objects}
\usage{
\method{nobs}{power_model}(object, ...)
}
\arguments{
\item{object}{A 'power_model' object.}

\item{...}{Unused.}
}
\value{
Number of observations used in the fit.
}
\description{
nobs method for power_model objects
}
