% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.power_model}
\alias{predict.power_model}
\title{Delta-method confidence intervals on power predictions}
\usage{
\method{predict}{power_model}(object, newdata, level = 0.95, use.robust = FALSE)
}
\arguments{
\item{object}{A 'power_model' fit.}

\item{newdata}{Data frame with m1, m2.}

\item{level}{Confidence level.}

\item{use.robust}{Use robust variance.}
}
\value{
Data frame with estimates and CIs.
}
\description{
Delta-method confidence intervals on power predictions
}
