% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{print.power_model}
\alias{print.power_model}
\title{Print method for power_model objects}
\usage{
\method{print}{power_model}(x, ...)
}
\arguments{
\item{x}{A 'power_model' object.}

\item{...}{Unused.}
}
\description{
Print method for power_model objects
}
