% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{AIC.power_model}
\alias{AIC.power_model}
\title{AIC method for power_model objects}
\usage{
\method{AIC}{power_model}(object, ...)
}
\arguments{
\item{object}{A 'power_model' object.}

\item{...}{Further arguments passed to methods.}
}
\value{
A numeric value with class "AIC".
}
\description{
AIC method for power_model objects
}
