% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powerFuns.R
\name{power_fun}
\alias{power_fun}
\title{Compute power for cross-classified LMM}
\usage{
power_fun(m1, m2, beta0, theta, r, alpha = 0.05, test = "onesided")
}
\arguments{
\item{m1}{Number of level-1 units per cluster.}

\item{m2}{Number of level-2 units per cluster.}

\item{beta0}{Critical z-value (e.g., -1.96), negative for left tail.}

\item{theta}{True effect size parameter.}

\item{r}{Intraclass correlation.}

\item{alpha}{Significance level.}

\item{test}{'onesided' or 'twosided'.}
}
\value{
Power (probability of rejecting null).
}
\description{
Calculates analytic power based on freely estimated intercept, adapted from
Jiang et al. (2024).
}
