% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.power_model}
\alias{predict.power_model}
\title{Predict power with Delta-method CIs}
\usage{
\method{predict}{power_model}(
  object,
  newdata = NULL,
  level = 0.95,
  use.robust = FALSE,
  transform = object$transform
)
}
\arguments{
\item{object}{A 'power_model' fit.}

\item{newdata}{Optional data.frame with columns m1, m2.
If NULL, the stored m1/m2 in the fitted object are used.}

\item{level}{Confidence level.}

\item{use.robust}{Use robust (sandwich) covariance.}

\item{transform}{Use the same transformation as during fitting by default.}
}
\value{
List with Power_hat, se, ci.
}
\description{
Predict power with Delta-method CIs
}
