% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{sim_data}
\alias{sim_data}
\title{Simulate binary outcomes under power model}
\usage{
sim_data(
  M1M2_grid,
  beta0_true,
  beta1_true,
  beta2_true,
  alpha = 0.05,
  seed = NULL,
  test = "onesided",
  parameterization = "natural"
)
}
\arguments{
\item{M1M2_grid}{Data frame with columns m1 and m2.}

\item{alpha}{Significance level.}

\item{seed}{Random seed for reproducibility.}

\item{test}{'onesided' or 'twosided'.}

\item{parameterization}{If 'natural', then beta1 and beta2 correspond to first and second random variance. Might be numerically unstable.}
}
\value{
Data frame with simulated y and true power.
}
\description{
Simulate binary outcomes under power model
}
